+++ app/test/e2e/subscriptions-find-all.spec.js
@@ -61,6 +61,43 @@ describe('Find all subscriptions tests', () => {
             response.body.links.should.have.property('first').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
             response.body.links.should.have.property('last').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
         });
+
+        it('Get all subscriptions with x-rw-domain header should be successful and use that header on the links on the response', async () => {
+            mockGetUserFromToken(ROLES.MICROSERVICE);
+
+            const response = await requester
+                .get(`/api/v1/subscriptions/find-all`)
+                .set('Authorization', `Bearer abcd`)
+                .set('x-rw-domain', `potato.com`);
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('prev').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('next').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('first').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('last').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+        });
+
+        it('Get all subscriptions with x-rw-domain and referer headers should be successful and use the x-rw-domain header on the links on the response', async () => {
+            mockGetUserFromToken(ROLES.MICROSERVICE);
+
+            const response = await requester
+                .get(`/api/v1/subscriptions/find-all`)
+                .set('Authorization', `Bearer abcd`)
+                .set('x-rw-domain', `potato.com`)
+                .set('referer', `https://tomato.com/get-me-all-the-data`);
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('prev').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('next').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('first').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+            response.body.links.should.have.property('last').and.equal('http://potato.com/v1/subscriptions/find-all?page[number]=1&page[size]=10');
+        });
     });
 
     it('Finding all subscriptions is only allowed when the request is performed by a micro service, failing with 401 Unauthorized otherwise', async () => {